Added module to make the class pure and independent
