+++ lib/player.rb
@@ -1,3 +1,4 @@
+require_relative '..\lib\module_ui.rb'
 class Player
   attr_reader :player_move
   attr_reader :name, :icon
@@ -6,19 +7,5 @@ class Player
     @icon = icon
     @player_move = []
   end
-
-
-  def move(board)
-    puts "#{name} Enter your move"
-    choice = gets.chomp.to_i
-    if (1..9).cover?(choice)
-      if board.place_taken?(choice-1)
-        puts 'Place taken try again'
-      else
-        choice
-      end
-    else
-      puts 'Enter a valid number'
-    end
-  end
+  include UserInterface
 end
